Beginning of Database Connections
A simple Connection and settings php document were created. The SQL file
is for the databases that I have created.
I have created 3 tables: The AcceptDeclineQuestions table, the Accepted
Stats table, and the Declined Stats table.

The Questions table contains all of the information about the question,
the text, etc.
The accept and decline tables are what the stats will be changed by for
each of the different stats. This will be whether the user accepts or
declines.
When the user decides, the question table will be updated with the
boolean saying the question has been answered or not.

Note: There is currently no set up with the database and the website at
the moment. Only a simple connection php file and the sql tables.