Backend Database Connection and PHP Calls
NOTE: index.html will no longer be used since we need it to be named
index.php in order to use the PHP calls for the database connection
backend.

Current functionality properly connects to SQL file, and does a basic
var_dump of all of the questions in the array for the table containing
all of the AccDecQuestions.